Allow for more flexible frontend and backend network selection.
This commit lays the groundwork for making the Ceph barclamp production
ready.  It does this by allowing the ceph roles to not rely on a single
network named Ceph, and instead allow the choice of networks to be fully
attribute driven.

For now, this code is still dependent on the names of the networks -- it
does not use Greg's categories and groups mechanism.